WS3 3.3: Extract REPLs to library crates
- Add rh-foundation::cli::repl (ReplSession, ReplInput) behind repl feature
  - ReplSession wraps rustyline DefaultEditor with history auto-save on drop
  - ReplInput enum abstracts Interrupted/Eof/Line variants
  - rustyline dep moves to rh-foundation (repl feature) from rh-cli

- Extract rh-cql::repl (repl feature)
  - CQL multi-line REPL with :commands (:quit, :help, :debug, :compact)
  - Uses ReplSession for readline/Ctrl-C/Ctrl-D handling

- Extract rh-fhirpath::repl (repl feature)
  - FHIRPath REPL with persistent history and .commands (.help, .data, .load, .quit)
  - Uses ReplSession::with_history for session persistence

- Extract rh-vcl::repl (repl feature)
  - VCL REPL with translate/explain modes and .commands
  - Uses ReplSession for readline loop

- Update rh-cli: enable repl features on rh-cql, rh-fhirpath, rh-vcl deps
  - Remove direct rustyline dep (moves to rh-foundation/repl)
  - CLI dispatch calls rh_cql::repl::run_repl, rh_fhirpath::repl::run_repl,
    rh_vcl::repl::run_repl directly (thin dispatch, no REPL logic in CLI)
  - CLI command files shrink to thin dispatch (<150 lines each for REPL code)
