+++ apps/rh-cli/Cargo.toml
@@ -28,17 +28,16 @@ anyhow = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 tokio = { workspace = true }
-rustyline = { workspace = true }
 glob = "0.3"
 
 # Local workspace crates
 rh-foundation = { path = "../../crates/rh-foundation", version = "0.2.0-beta.2", features = ["http"] }
 rh-codegen = { path = "../../crates/rh-codegen", version = "0.2.0-beta.2" }
-rh-cql = { path = "../../crates/rh-cql", version = "0.2.0-beta.2" }
-rh-fhirpath = { path = "../../crates/rh-fhirpath", version = "0.2.0-beta.2" }
+rh-cql = { path = "../../crates/rh-cql", version = "0.2.0-beta.2", features = ["repl"] }
+rh-fhirpath = { path = "../../crates/rh-fhirpath", version = "0.2.0-beta.2", features = ["repl"] }
 rh-packager = { path = "../../crates/rh-packager", version = "0.2.0-beta.2" }
 rh-validator = { path = "../../crates/rh-validator", version = "0.2.0-beta.2" }
-rh-vcl = { path = "../../crates/rh-vcl", version = "0.2.0-beta.2" }
+rh-vcl = { path = "../../crates/rh-vcl", version = "0.2.0-beta.2", features = ["repl"] }
 rh-hl7-fhir-r4-core = { path = "../../crates/rh-hl7_fhir_r4_core", version = "0.2.0-beta.2" }
 rh-fsh = { path = "../../crates/rh-fsh", version = "0.2.0-beta.2" }
