+++ crates/rh-cql/Cargo.toml
@@ -44,6 +44,7 @@ path = "src/bin/test_parse.rs"
 
 [features]
 default = []
+repl = ["rh-foundation/repl"]
 
 [lints]
 workspace = true
