+++ crates/rh-fhirpath/Cargo.toml
@@ -17,6 +17,7 @@ crate-type = ["rlib"]
 
 [features]
 wasm = ["dep:wasm-bindgen", "dep:js-sys", "dep:console_error_panic_hook"]
+repl = ["rh-foundation/repl"]
 
 [package.metadata.wasm-pack.profile.release]
 wasm-opt = ["-O4"]
