+++ crates/rh-foundation/Cargo.toml
@@ -24,6 +24,7 @@ tokio = { workspace = true, optional = true }
 url = { workspace = true, optional = true }
 tar = { workspace = true, optional = true }
 flate2 = { workspace = true, optional = true }
+rustyline = { workspace = true, optional = true }
 wasm-bindgen = { version = "0.2", optional = true }
 console_error_panic_hook = { version = "0.1", optional = true }
 
@@ -39,6 +40,7 @@ harness = false
 [features]
 default = []
 http = ["reqwest", "tokio", "url", "tar", "flate2"]
+repl = ["rustyline"]
 wasm = ["wasm-bindgen", "console_error_panic_hook"]
 
 [lints]
