+++ crates/rh-vcl/Cargo.toml
@@ -18,6 +18,7 @@ crate-type = ["rlib"]
 
 [features]
 wasm = ["dep:wasm-bindgen", "dep:js-sys", "dep:console_error_panic_hook"]
+repl = ["rh-foundation/repl"]
 
 [package.metadata.wasm-pack.profile.release]
 wasm-opt = ["-O4"]
